DCE-28: Update Pawn Attacks
