+++ src/engine/attacks/mod.rs
@@ -1,3 +1,8 @@
-pub mod all_attacks;
-pub mod pawn_attacks;
-pub mod pext;
+pub mod bishop;
+pub mod king;
+pub mod knight;
+pub mod pawn;
+pub mod queen;
+pub mod rays;
+pub mod rook;
+pub mod utility;
