+++ src/lib.rs
@@ -1,9 +1,14 @@
 pub mod engine {
     pub mod game;
     pub mod attacks {
-        pub mod all_attacks;
-        pub mod pawn_attacks;
-        pub mod pext;
+        pub mod bishop;
+        pub mod king;
+        pub mod knight;
+        pub mod pawn;
+        pub mod queen;
+        pub mod rays;
+        pub mod rook;
+        pub mod utility;
     }
     pub mod move_generation {
         pub mod fen;
